Relocate Lit peek impls into lit module
